V1.1
World Diameter wasn't being calculated correctly (in worldDiameter() in
SystemSupport.h), it was allowing ranges outside of it its range. It was
sloppy math
Corrected the spelling of marginality (was Magrinality...) throughout
the comments and titles
Fixed some issues of code ignorning forbidden zones and generating
planets where there shouldn't have been
The forbidden range was calculated and forbidden flags were set, but
none of it was referenced again
Fixed an issue where planets weren't being generated but should have
been
Primary star has a forbidden radius of 0.0, which is always less than
the inner limit radius, therefore, it doesn't get planets
Changed how planets align in  their arrays, they now align on the 0
instead of the 1
This means all the loops that assumed 1 needed updating (and they are
all pervading)
This was a kludge at the time, but I ignored early in the project. Later
when it was understood what was happening, I could have fixed it (but
that would have been a lot of work, so I had left it as is)
RVM was modified to no longer factor in tectonics when determining an
RVM score (Was minorily affecting the output when it shouldn't have)
Changed how files were opened/closed giving what looks like a 3x speed
boost (opens before and after all generation is done, not both on each
record)
In addition to this, I was forgetting to close the main file, so it was
never being closed, though constantly being opened to write to
This meant that corruption was bound to happen, and this was the cause
of the half written records that were overlapping the previous record